+++ resources/js/metabox-validation.js
@@ -1,0 +1,5 @@
+jQuery(document).ready(function(){
+    
+    // TODO: Validate what needs validation here!
+    
+});